moving the getShocks() test into ConsumptionSaving/tests/
